+++ netunicorn-base/src/netunicorn/director/base/connectors/types.py
@@ -1,0 +1,6 @@
+from typing import TypedDict
+
+
+class StopExecutorRequest(TypedDict):
+    executor_id: str
+    node_name: str
